ajout check_textures.c + norme textures 01/04
