+++ only_colors.cub
@@ -1,4 +1,4 @@
-NO ./textures/north.xpm
+NO textures/north.xpm
 SO              ./textures/south.xpm
 EA       ./textures/north.xpm
 WE      ./textures/west.xpm
