+++ srcs/parsing/textures_and_colors_parser/check_colors.c
@@ -1,0 +1,13 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   check_colors.c                                     :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: racoutte <racoutte@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/04/01 16:43:03 by racoutte          #+#    #+#             */
+/*   Updated: 2025/04/01 16:49:34 by racoutte         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub_parsing.h"
