+++ srcs/parsing/textures_and_colors_parser/textures_and_colors.c
@@ -6,41 +6,15 @@
 /*   By: racoutte <racoutte@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/28 12:37:57 by racoutte          #+#    #+#             */
-/*   Updated: 2025/04/01 14:41:25 by racoutte         ###   ########.fr       */
+/*   Updated: 2025/04/01 16:49:14 by racoutte         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub_parsing.h"
 
-static bool	is_texture_missing(t_textures textures)
-{
-	if (textures.east_texture == NULL || textures.west_texture == NULL)
-		return (true);
-	if (textures.north_texture == NULL || textures.south_texture == NULL)
-		return (true);
-	return (false);
-}
-
 static void	check_textures_and_colors_are_valid(t_game_data *game_data)
 {
-	if (is_texture_missing(game_data->textures) == true)
-	{
-		if (game_data->parse_error_info.error_type != NO_ERROR)
-		{
-			ft_dprintf(STDERR_FILENO, "Error\n%s: invalid texture.\n",
-				game_data->parse_error_info.invalid_element);
-		}
-		else
-			ft_dprintf(STDERR_FILENO, "Error\nA texture path is missing.\n");
-		parser_exit_routine(game_data);
-		exit(FAILURE);
-	}
-	if (is_double_texture(game_data) == true)
-	{
-		ft_dprintf(STDERR_FILENO, "Error\nDouble texture detected.\n");
-		parser_exit_routine(game_data);
-		exit(FAILURE);
-	}
+	check_textures(game_data);
 }
 
 void	get_textures_and_colors(t_game_data *game_data)
@@ -58,3 +32,6 @@ void	get_textures_and_colors(t_game_data *game_data)
 
 //return (fonction de fin de processus); fonction de fin fait son taff
 //et retourne un t_content_status selon ce qu'elle a trouve
+
+//check_colors(game_data); en 2eme ligne de
+//check_textures_and_colors_are_valid
